Added a group_indices_based_on_label! method to Sum instances
This introduces the class of DistributedIndex which represents an index
in the overall index structure of a sum. It holds references to the indices
from the terms in the sum which contribute to the overall index. Some more
work is needed to handle contractions correctly and make indices in sums
aware that they are contracted through a DistributedIndex