Refactored the way DistributedIndex instances are extracted from sums, and how the options for dealing with indices are passed.
There is now a global set of properties which control how expressions are interpreted. These can also be overridden down the chain as the expression is built up. DistributedIndex instances are automatically created as terms are added into a sum and so the need for a sanitize! method is removed. The expression should never be in an inconsistent state right after being parsed now.