Seperated object initialization into two distinct parts: allocation of the object, and setting up of the object properties
The new/initialize method is now only responsible for directly assigning instance methods. This should only be called from other parts of the program which know what they are doing since the object can easily be put in an inconsistent state. The code for assigning the correct properties to an expression and making it consistent (e.g. assigning index#position_matters, or extracting DistributedIndex instances) has been moved to a method called apply_properties!. This is called from the parser with the properties passed to the parser as the expression is created. This new set up allows objects to be copied and recreated by directly assigning the correct instance methods, but without inadvertently calling code which tried to be clever and modify them.