Added #new_from_serialization and other helper methods to build up an expression from a serialization.
This will build up an expession in the same state as the one that created the serialization.
This has the caveat that the initial serialization ignores external contractions and so the
resulting expression will only be equivalent up to external contractions. eql? and == also ignore
contractions so they will still return the two expressions being equivalent.