Added a Wrapper class around all Expressions that are built.
This is to allow us to do in place modifications of expressions. Performing an
operation on an expression may change the type of expression that it is. E.g.
split_indices sends Product -> Sum. Since we can't change the class of an Expression
we need to wrap everything so that we can.